+++ tests/test_cmd.py
@@ -423,30 +423,24 @@ def test_cmdoutput_object_with_non_subprocess_backends():
 
     # In this case, when tee=0 the user can still capture the output
     info = ub.cmd('echo hello world', detach=True, capture=True, tee=0)
-    with pytest.raises(KeyError):
-        info.stdout
-    with pytest.raises(KeyError):
-        info.stderr
+    assert info.stdout is None
+    assert info.stderr is None
     assert info['proc'].communicate()[0] is not None
 
     # In this case, when tee=0 and capture=False, the user cannot capture the output
     info = ub.cmd('echo hello world', detach=True, capture=False, tee=0)
-    with pytest.raises(KeyError):
-        info.stdout
-    with pytest.raises(KeyError):
-        info.stderr
+    assert info.stdout is None
+    assert info.stderr is None
     assert info['proc'].communicate()[0] is None
 
     # In this case when tee=1, a detached process will show its output but
     # capturing will not be possible.
     info = ub.cmd('echo hello world', detach=True, capture=False, tee=1)
-    with pytest.raises(KeyError):
-        info.stdout
-    with pytest.raises(KeyError):
-        info.stderr
+    assert info.stdout is None
+    assert info.stderr is None
     info['proc'].communicate()
     info.args
-    with pytest.raises(KeyError):
+    with pytest.raises(RuntimeError):
         info.check_returncode()
 
     # Check attributes when system=True